+++ client/src/services/ApiClient.js
@@ -18,3 +18,18 @@ export async function getEmployeeByID(employeeID) {
       throw new Error("An error occurred during employee fetch");
   }
 }
+
+// --------
+// Order
+// --------
+
+export async function createOrder(clientId, products) {
+  var obj = { clientId: clientId, products: products };
+  console.log(obj);
+
+  await fetch("/api/orders", {
+    method: "POST",
+    headers: { "Content-Type": "application/json" },
+    body: JSON.stringify({ ...obj }),
+  });
+}
