+++ client/src/ui-components/ShoppingCartComponent/ShoppingCartControlsCSS.css
@@ -14,6 +14,14 @@
   margin-left: 1%;
   border-color: #db9471;
 }
+.cart-inverse-button:focus {
+  border: 2px solid black;
+  background-color: #ece6e2;
+  color: #db9471;
+  width: auto;
+  margin-left: 1%;
+  border-color: #db9471;
+}
 .cart-button {
   border: 2px solid black;
   background-color: #db9471;
@@ -21,6 +29,12 @@
   border-color: #db9471;
 }
 .cart-button:hover {
+  border: 2px solid black;
+  background-color: #d88054;
+  width: auto;
+  border-color: #db9471;
+}
+.cart-button:focus {
   border: 2px solid black;
   background-color: #d88054;
   width: auto;
