+++ client/src/ui-components/ShoppingCartComponent/ShoppingCartTotAmount.js
@@ -1,14 +1,16 @@
-import 'bootstrap/dist/css/bootstrap.min.css';
-import './ShoppingCartTotAmountCSS.css';
-import React from 'react';
+import "bootstrap/dist/css/bootstrap.min.css";
+import "./ShoppingCartTotAmountCSS.css";
+import React from "react";
 
 function ShoppingCartTotAmount(props) {
-
-    return (
-        <div className="total-amount">
-            <h4>Total amount: {props.tot.toFixed(2)}</h4>
-        </div>
-    );
+  return (
+    <div className="total-amount">
+      <h4>
+        Total amount: {props.tot.toFixed(2)}
+        {" €"}
+      </h4>
+    </div>
+  );
 }
 
 export { ShoppingCartTotAmount };
